move api methods to helper class
